Move EditorConfigOptionsEnumerator to Features
