+++ src/Features/Core/Portable/Options/EditorConfig/EditorConfigOptionsEnumerator.cs
