+++ src/Features/Core/Portable/Options/EditorConfig/EditorConfigOptionsEnumeratorAttribute.cs
