+++ src/Features/Core/Portable/Options/EditorConfig/IEditorConfigOptionsEnumerator.cs
